requirements.txt
Flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.32.2
stripe==9.4.0
postmarker==1.0
